refactor(utils/db): tidy up db utils into model resource structure
