+++ app/models/animal/getAnimalsByFilter/constants.ts
