+++ app/routes/favorites/index.tsx
@@ -1,7 +1,7 @@
 import type { ActionFunction } from '@remix-run/node';
-import type { AnimalId } from '~/utils/db/getAnimalByIds';
+import type { AnimalId } from '~/models/animal/getAnimalsByIds/index.server';
 import { json } from '@remix-run/node';
-import getAnimalByIds from '~/utils/db/getAnimalByIds';
+import getAnimalByIds from '~/models/animal/getAnimalsByIds/index.server';
 import parsePayloadByJson from '~/utils/action/parsePayloadByFormData';
 import Favorites from '~/features/favorites';
 import Layout from '~/components/common/Layout';
