+++ app/routes/index.tsx
@@ -1,10 +1,10 @@
 import type { ActionFunction } from '@remix-run/node';
-import type { Filter } from '~/utils/db/getAnimalsByFilter';
+import type { Filter } from '~/models/animal/getAnimalsByFilter/index.server';
 import { useEffect, useState } from 'react';
 import PetPairing from '~/features/pairing';
 import FullPageLoading from '~/components/common/FullPageLoading';
 import Layout from '~/components/common/Layout';
-import getAnimalsByFilter from '~/utils/db/getAnimalsByFilter';
+import getAnimalsByFilter from '~/models/animal/getAnimalsByFilter/index.server';
 import parsePayloadByFormData from '~/utils/action/parsePayloadByFormData';
 import { json } from '@remix-run/node';
